cloud_functions/retraining_notification/main.py
from google.cloud import aiplatform

def trigger_retraining(event, context):
    # Set your project and location
    project = "your-project-id"
    location = "us-central1"

    # Initialize the Vertex AI client
    aiplatform.init(project=project, location=location)

    # Trigger a pipeline or custom training job
    pipeline = aiplatform.PipelineJob(
        display_name="retraining-pipeline",
        template_path="gs://your-template-path",
        parameter_values={
            "input_data": "gs://your-new-training-data-path",
            # Add any additional parameters
        }
    )
    pipeline.run(sync=False)