cloud_functions/retraining_notification/requirements.txt
google-cloud-aiplatform
google-cloud-storage
functions-framework
